Update cell formatting in MemoryPrinter
Adjusted the formatting of the "formattedLine" string in MemoryPrinter to accommodate for variability in name length. Modified the SetInitModeCommand to initialize with the default if the mode is set to STOP_MODE_SYNTAX and meets the lower limit number of arguments. These changes ensure more robust and flexible representation and initialization of memory cells.