Refactor AI memory allocation and handling
The changes include introducing dynamic memory allocation for each AI during memory loading phase in the CodeFight module. This is achieved by including memory allocation property in the AI class. Also loaded memory navigation is altered in the game handler. These changes improve the handling and usage of memory in the simulation.